Add clipper-gemmi library to the build process
The build process now includes a new library, clipper-gemmi. This is reflected in updates to the 'CMakeLists.txt' files, specifying its inclusion in the subdirectories and linking it to the 'moorhen' executable. A new 'CMakeLists.txt' file for the 'clipper-gemmi' project was also created, outlining its build and installation instructions.