Introduced Carbohydrate Section in Model Menu
This commit involves a refactor improving the handling of Privateer validation results in coot/moorhen-wrappers.cc. In addition, a new MoorhenCarbohydrateCard was created which contains the code to render a set of cards each containing a single carbohydrate SNFG. A section in the model menu was added to display this information, and the previously developed Carbohydrate validation section was refactored to use this base card.