Update Privateer library version and refactor wrappers code
The changes include updating the version of the Privateer library being used and moving the wrapper functionality related to Privateer into a new header file. This segregation of functions into a dedicated header file provides a cleaner codebase. The unnecessary console outputs in the Privateer wrapper functions have been commented out, reducing the verbosity of the code.